pyproject.toml
[project]
name = "reconciliation-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.45",
    "fabric>=3.2.2",
    "fastexcel>=0.12.0",
    "pandera[polars]>=0.20.4",
    "polars-lts-cpu>=1.9.0",
]

[tool.uv]
dev-dependencies = [
    "jupyterlab-vim==4.1.0",
    "jupyterlab==4.0.9",
    "pytest>=8.3.3",
]
